Disable debug assertions
Since Rust 1.78, there are debug that should detect undefined behaviour. Unfortunately, on Android there seems to be some UB, so the app instantly crashes when running in debug mode. However, testing the android app in release mode would be a pain (one has to sign a release built app with their key), so this is the best way I can currently think of. Debug assertions are only disabled in crates, so our code is checked. They are also enabled for the ring crate since it won't build without them.