Save and restore the actual theme
This way the theme is saved so that the customizations are not lost
Currently the " (restored)" suffix is added if the theme does not match a theme with the same name. Not sure if we want to check for any matching themes and switch to those? Could be confusing I suppose...